# backend/public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion des Quotes-Parts - Authentification",
        "description": "Documentation des routes d’authentification (login, logout, me).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion d’un utilisateur",
                "description": "Permet de se connecter avec un nom d'utilisateur et un mot de passe.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "properties": {
                                                "UTI_CODE": {
                                                    "type": "string"
                                                },
                                                "UTI_NOM": {
                                                    "type": "string"
                                                },
                                                "UTI_USERNAME": {
                                                    "type": "string"
                                                },
                                                "UTI_STATUT": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants incorrects"
                    },
                    "403": {
                        "description": "Compte inactif"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Obtenir les informations de l’utilisateur connecté",
                "description": "Retourne les informations du compte actuellement connecté.",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "Informations de l’utilisateur connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "UTI_CODE": {
                                            "type": "string"
                                        },
                                        "UTI_NOM": {
                                            "type": "string"
                                        },
                                        "UTI_USERNAME": {
                                            "type": "string"
                                        },
                                        "UTI_STATUT": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion d’un utilisateur",
                "description": "Révoque le token d’accès actuel de l’utilisateur connecté.",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user-fonctionnalites": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Obtenir les fonctionnalités de l’utilisateur connecté",
                "description": "Retourne la liste des codes de fonctionnalités associées au groupe de l'utilisateur connecté.",
                "operationId": "9e023d757130ca1dc708e80377fa304e",
                "responses": {
                    "200": {
                        "description": "Liste des fonctionnalités récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}